poc_face_recognition_and_classification/requirements.txt
certifi==2020.6.20
click==7.1.2
cmake==3.18.2.post1
dlib==19.21.0
face-recognition==1.3.0
face-recognition-models==0.3.0
numpy==1.19.4
Pillow==8.0.1
wincertstore==0.2
tensorflow==2.2.0
keras==2.3.1
opencv-python==4.2.0.*
imutils==0.5.3
matplotlib==3.3.2
sklearn